initial commit for nano s - derived from ledger-app-neo - adapted to cpx - supports get pubkey - supports signing messages
