+++ .project
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <projectDescription>
-	<name>blue-app-neo</name>
+	<name>ledger-app-cpx</name>
 	<comment></comment>
 	<projects>
 	</projects>
